Password bubble: Don't focus by default.
This is the duplicate of https://codereview.chromium.org/373043003. The bubble turned to be unresponsive on Windows. This CL has a workaround. The original problem is in infrastructure. It's to be fixed in  BubbleDelegateView.
set_use_focusless(true) has adverse effect on Windows.
1) The bubble can't be the active window (which is not desired behavior).
2) The bubble doesn't get WM_LBUTTONDOWN event. Therefore, all the buttons get stuck.
We call set_use_focusless(false) after the bubble is shown. Thus, it can get focus and mouse events properly.

BUG=383821

Review URL: https://codereview.chromium.org/396973002