Fix EnterpriseDeviceManagementStatus UMA enum, add comments.
I've gone through the enterprise UMA enums and compared them to their C++
counterparts.  There was a mismatch for EnterpriseDeviceManagementStatus which
is corrected by this CL.  To prevent this kind of error in the future, this CL
adds comments and numeric ids to the C++ enums.

BUG=none

Review URL: https://codereview.chromium.org/375903002