Docserver: Modify doc inlining so all relevant data is present in availability_finder.py
Not inlining schema nodes for availability finder causes data to go missing and
availability to not be found. Inlining schema nodes means that availability data
gets messed up when a node is no longer inlined, because inlined nodes are
deleted from the schema. Solution: Inline nodes, but do not remove them from the
schema when used in availability_finder.py. This provides all the necessary data,
and doesn't interfere with availability data generation when a node is no longer
inlined because it will be still be present in the schema graphs.

BUG=233982
NOTRY=True

Review URL: https://codereview.chromium.org/385703005