Makes it so a node can only the root of one connection at a time
This resulted in needing to make structural changes. For example, if
connection 2 has a root of A, and A has the children B and C then if A
is assigned to another connection (by way of the owner calling Embed
again), then connection 2 is told A has been deleted and B and C are
removed as children of A.

Because this does structural changes the server change id is advanced.

BUG=389339
TEST=covered by unit tests
R=ben@chromium.org

Review URL: https://codereview.chromium.org/397803003