GLHelper: validate destination size matches last scale size
In the GLHelper unittest the ValidateScalerStages function checks that
the scaling stages to be executed for the scale are valid.

This change also adds a check to ensure that the size of the destination
texture matches the destination size in the last stage of the scaler
pipeline.

This identified some of the test cases had small errors in the
destination sizes which didn't match the description of the scale. The
patch also corrects those.

TEST=content_gl_tests passes
BUG=None

Review URL: https://codereview.chromium.org/394203002