Simplify the logic in DesktopWindowTreeHostX11 for FocusOut
We should have the same handling regardless of whether the 'mode' member is
NotifyGrab or NotifyUngrab. According to
http://tronche.com/gui/x/xlib/events/input-focus/grab.html, whether the mode
field is NotifyGrab or NotifyUngrab depends on whether there is a new
capturer for the keyboard. The X window has lost keyboard capture in both cases.

BUG=392480
TEST=None

Review URL: https://codereview.chromium.org/379773003