NaCl SDK: Demonstrate use of pp::Graphics2D::SetScale()
Demonstrate in the NaCl SDK API demo for pp::Graphics2D how
to use full device resolution on HiDPI displays. For Graphics2D,
the example now calls pp::View::GetDeviceScale(), and creates
a Graphics 2D context with the correct number of pixels, as well
as show when to call pp::Graphics2D::SetScale().

It also scales the mouse position to the correct location and
increases the radius of the circle created in DrawMouse() to match.

BUG=144071,392557
TEST=Build NaCl SDK, run with --force-device-scale-factor=2

Review URL: https://codereview.chromium.org/397173003