Fix flipped views on Mac
Setting a layer-hosting NSView's CALayer's geometry flipping has a lot
of unexpected side-effects, and is proving unsafe, especially if that
NSView will have sub-views.

Add a level of indirection where the hosted layer is not flipped, but it
has a sub-layer that is flipped, off of which the layers for composited
content are attached.

BUG=392952

Review URL: https://codereview.chromium.org/392143002