Fix canPlayType() support for non-RFC compliant mp3 mimetype.
The 'audio/mpeg; codecs="mp3"' mimetype is not RFC 3003 compliant, but a
bunch of sites apparently use it. This change restores the behavior that
was in M35 and earlier and returns "probably" for this mimetype.

The changes are intentionally kept small so they can be easily be merged
to the M36 & M37 branches.

BUG=393720,386073
TEST=MediaCanPlayTypeTest.CodecSupportTest_mp3

Review URL: https://codereview.chromium.org/400023002