Cleanup: remove IdentityAccountTracker (replaced by gaia::AccountTracker)
This change replaces the IdentityAccountTracker with
gaia::AccountTracker, which mostly does the same thing. The one
functional consequence of this change is that the Identity API no
longer keeps track of login token errors on its own. Instead it
relies on the underlying token services to report errors to the
user.

BUG=364203

Review URL: https://codereview.chromium.org/398573007