Add the ability for presentation timestamps to be specified in test strings.
All SourceBufferStream string based unit tests assume that decode and
presentation timestamps are the same. This change extends the string
syntax so that buffers can be created where the presentation and
decode timestamps are different. This will make it easier to write
tests that verify proper behavior when B-frames are used.

Review URL: https://codereview.chromium.org/395303002