Check isActive in onRestore to work-around crbug.com/394819.
Also fixes a related bug involving exiting full-screen. If the window was previously maximized when it is full-screened, then ChromeOS sometimes generates an onMaximized event when exiting full-screen, which triggers the app to re-enter full-screen mode. The fix is to explicitly unmaximize the window before going full-screen (which also has the nice benefit of making the double restore() calls unnecessary).

BUG=384678

Review URL: https://codereview.chromium.org/398403002