Enable x64 Chromoting host and installer
N.B. I had to move the sawbuck_provider in the .wxs out of the
installation directory. Even though it only contains registry changes,
candle will complain if it is not inside a directory, but the
application directory is platform specific, whereas sawbuck_provider is
always 32-bit.

BUG=394467

Review URL: https://codereview.chromium.org/398173002