[Profiles] Create a profile moar better after deleting the only existing profile.
The problem was that we only created the new profile if we deleted the
active profile, and it was the only existing profile. Since guest mode came
along, we could also delete the only profile when the active profile was
the guest one (which doesn't really count as a profile), so you'd be left in
this borked state of "no profiles in Chrome, oops".

For brownie points, I also added tests for both of these "only existing profile
deleted" scenarios. :)

BUG=393901
TEST=See bug description.

Review URL: https://codereview.chromium.org/396773002