Set PPAPI plugins' requests' contexts to 'WebURLRequest::RequestContextPlugin'.
Pepper plugins' requests are piped through Blink's Fetch implementation,
giving us the opportunity to make policy decisions about whether or not
they should be allowed by CSP or Mixed Content (we're currently not
using this ability).

In order to make these decisions, we need to know that the request is
for data a plugin is requesting, as opposed to a plugin itself. This
patch adds that detail to PPAPI's URL-requesting logic.

BUG=390497

Review URL: https://codereview.chromium.org/409563002