Make the Win-Clang Dbg FYI builder do a shared_library build
This is the default for Debug builds. We were previously forcing it
to do a static build because Clang didn't have good enough dllimport/export
support, but that should work now.

The shared library build doesn't work with NaCl yet, so disable it for now.

BUG=82385

Review URL: https://codereview.chromium.org/407003003