Use a correct scale factor in app_list::CachedLabel
It currently just uses 1.0. Instead, fetch it with
ui::GetScaleFactorForNativeView and invalidate it when the scale factor
changes.

This is a small fix for a simple HiDPI bug, but eventually we will be
able to scrap CachedLabel: The performance improvement it offers will be
redundant with http://crbug.com/240037.

BUG=368493

Review URL: https://codereview.chromium.org/405373002