VaapiVEA: Handle bitrate and frame rate negotiation from webrtc
The webrtc video encoder path will call into RTCVideoEncode::SetRates
with a zero framerate as part of the video encode setup.

A zero frame rate is not accepted by the VAAPI encoder and so we need to
bump the framerate up to a minimum of 1.

This change mimics the behavior of the V4L2 based encoder which does the
same thing to the framerate and bitrate.

TEST=Run chrome --enable-vaapi-accelerated-video-encode and use the
Google Cast extension to share the tab to the Chrome Cast.
BUG=391445

Review URL: https://codereview.chromium.org/369923008