Simplify RunTaskOnThread
RunTaskOnThread has been running tasks synchronously if the task runner belongs to the current thread.
This behavior was necessary when it was allowed to access drive::FileSystem from both UI and IO threads.
But now, no one needs this behavior and every user uses this method to post tasks to the specified runner.

BUG=None

Review URL: https://codereview.chromium.org/408153003