Mac: Fix bad framerate when capturing tab
The fix for this issue, namely, that we must draw contents manually and
immediately when a window is occluded because Cocoa will not draw them
for us, was not ported to the browser compositor.

Plumb this explicitly through a ShouldAckImmediately callback for
CompositingIOSurfaceLayerClient and BrowserCompositorViewMacClient.

Also check for vsync disabling in the ShouldAckImmediately instead of
running it (unnecessarily) through CompositingIOSurfaceContextMac.

BUG=393390,350410

Review URL: https://codereview.chromium.org/408103004