content: Add DeletedGpuMemoryBuffer IPC.
This IPC is used by the renderer to notify the browser when
a GPU memory buffer instance is deleted. This allows the browser
to maintain proper lifetime management of GPU memory buffers for
which ownership can't be passed with IPC.

This also renames the GpuCommandBufferMsg_DestroyGpuMemoryBuffer
IPC to GpuCommandBufferMsg_UnregisterGpuMemoryBuffer to not
be confused with real GPU memory buffer destruction, which
doesn't necessarily happen at when the GLImage is destroyed.

BUG=

Review URL: https://codereview.chromium.org/409723003