Add support to send print preset options in one message
This patch introduces structure that should be filled with all
print preset options that we need to get from PDF and set in print
preview.

Before this patch one message was being sent to browser to set PrintScaling
property but going forward, more properties will be supported and so
it is better to send only one message with all properties set in one structure.

BUG=169120

Review URL: https://codereview.chromium.org/388783003